feat(collector): add ratelimit source label to metrics
This adds a `source_ip` label to the two existing metrics in order to
differentiate which egress IP a ratelimit pertains to.

In our use-case, we want to run `docker-hub-rate-limit-exporter` as a
daemonset on nodes that run across AZ boundaries. In this case, each AZ
has its own egress IP, and therefore, it's own ratelimit.

With the status quo, if I am to run this exporter across two AZs, such
as us-east-1a and us-east-1b, there's no way to differentiate which pod
is reporting for which egress ip. If one AZ's limit starts being
exhausted, and I were graphing the results, it would also be very
confusing, as one would be reporting 100, while the other would start
reporting, say, 60.

Adding the source ip as an additional label should not affect existing
metrics and dashboards, but gives another dimension so that we can
differentiate which ratelimit is which within our environments.